+++ dart/geobase/lib/geobase.dart
@@ -27,6 +27,7 @@ export 'src/codes/axis_order.dart';
 export 'src/codes/canvas_origin.dart';
 export 'src/codes/cardinal_precision.dart';
 export 'src/codes/coords.dart';
+export 'src/codes/dimensionality.dart';
 export 'src/codes/dms_type.dart';
 export 'src/codes/geo_representation.dart';
 export 'src/codes/geom.dart';